backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "psycopg2>=2.9.10",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]
